Time: 27 ms (56.97%), Space: 48.9 MB (6.08%) - LeetHub

+++ 1283-find-the-smallest-divisor-given-a-threshold/1283-find-the-smallest-divisor-given-a-threshold.java
@@ -24,9 +24,9 @@ class Solution {
 
     }
 
-    public static int check(int[] arr,int mid)
+    public static double check(int[] arr,int mid)
     {
-        int sum=0;
+        double sum=0;
         for(int i=0;i<arr.length;i++)
         {
             sum+=Math.ceil((double)arr[i]/mid);
